+++ src/bittytax/conv/parsers/__init__.py
@@ -45,6 +45,7 @@ from . import (
     hitbtc,
     hotbit,
     ii,
+    kinesis,
     koinly,
     kraken,
     kucoin,